🚚 Move responses directory to test/types
