+++ tests/types/responses/get-data.ts
@@ -1,9 +1,9 @@
 import { config } from 'dotenv'
 config()
-import authorize from '../../authorize'
-import * as requests from '../../src/requests'
+import authorize from '../../../authorize'
+import * as requests from '../../../src/requests'
 import { writeFileSync } from 'fs'
-import { params } from '../global'
+import { params } from '../../global'
 import { dataPath } from '.'
 
 async function getData() {
